Change tear down from delete to drop tables
