+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\__init__.py":[1,2,3,5,6,7,8,9,10],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\redflag_views.py":[1,2,3,5,7,10,11,16,17,22,23,28,33,38,43],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\redflag_controllers.py":[1,2,3,4,7,9,12,53,60,67,79,101,117,10],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\redflag_models.py":[1,2,5,6,9,20,27,35,43,50,57],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\database.py":[1,2,3,5,6,8,43,90,10,11,12,13,14,25,26,27,28,29,30,31,33,34,35,37,38,56,70,83,85,86,87,88,93,94,95,96],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\user_views.py":[1,2,3,5,7,10,15,20,25,12,17],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\user_controllers.py":[1,2,3,4,7,10,12,15,93,117,120,13,16,18,20,21,22,23,24,25,26,27,29,35,41,46,51,56,57,62,63,69,70,71,72,73,74,75,77,79,80,82,84,85,86,88,89,91,95,97,99,101,113,114,115,58,59,60,61,103,104,106,108,109,110,111,112],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\user_models.py":[1,2,3,5,6,9,22,30,38,44,53,7,32,33,34,35,36,24,25,26,27,28,11,14,16,17,18,19,20,40,41,42,43],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\intervention_views.py":[1,2,3,4,6,8,11,16,17,22,26,27,31,32,36,37,41,42,46,47,52,13,19],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\intervention_controllers.py":[1,2,3,4,7,9,12,53,60,68,75,87,109,125,10,14,15,16,17,18,19,20,22,23,25,26,31,32,38,39,40,41,42,45,46,48,49,51],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\intervention_models.py":[1,2,5,6,9,20,27,35,43,51,58,65,7,10,14,15,16,17,18],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\utilities\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\utilities\\auth.py":[1,2,3,4,7],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\test_endpoints.py":[1,2,5,7,15,33,53,73,101,116,8,9,10,12,13,55,56,57,58,59,60,61,63,64,66,67,69,70,71,102,104,105,106,107,109,110,111,112,113,114,117,119,120,121,122,124,125,126,127,128,18,19,20,21,22,23,24,26,27,28,29,30,31,75,76,77,78,79,80,81,83,84,86,87,88,89,90,91,92,94,95,96,97,98,99,35,36,37,38,39,40,41,43,44,46,47,49,50,51],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\base.py":[1,2,3,4,7,8,13,16,31,9,10,11,14,17,18,19,20,21,22,23,25,26,27,28,29]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\__init__.py":[1,2,3,5,6,7,8,9,10],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\redflag_views.py":[1,2,3,5,7,10,11,16,17,22,23,28,33,38,43],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\redflag_controllers.py":[1,2,3,4,7,9,12,53,60,67,79,101,117,10],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\redflag_models.py":[1,2,5,6,9,20,27,35,43,50,57],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\database.py":[1,2,3,5,6,8,43,90,10,11,12,13,14,25,26,27,28,29,30,31,33,34,35,37,38,56,70,83,85,86,87,88,93,94,95,96],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\user_views.py":[1,2,3,5,7,10,15,20,25,12,17],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\user_controllers.py":[1,2,3,4,7,10,12,15,93,117,120,13,16,18,20,21,22,23,24,25,26,27,29,35,41,46,51,56,57,62,63,69,70,71,72,73,74,75,77,79,80,82,84,85,86,88,89,91,95,97,99,101,113,114,115,58,59,60,61,103,104,106,108,109,110,111,112],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\user_models.py":[1,2,3,5,6,9,22,30,38,44,53,7,32,33,34,35,36,24,25,26,27,28,11,14,16,17,18,19,20,40,41,42,43],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\views\\intervention_views.py":[1,2,3,4,6,8,11,16,17,22,26,27,31,32,36,37,41,42,46,47,52,19,24,34,13],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\controllers\\intervention_controllers.py":[1,2,3,4,7,9,12,53,60,68,75,87,109,125,10,14,15,16,17,18,19,20,22,23,25,26,31,32,38,39,40,41,42,45,46,48,49,51,54,55,56,57,69,70,71,72],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\models\\intervention_models.py":[1,2,5,6,9,20,27,35,43,51,58,65,7,10,14,15,16,17,18,22,23,24,25,37,38,39,40,41],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\utilities\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\utilities\\auth.py":[1,2,3,4,7],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\__init__.py":[1],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\test_interventions.py":[1,2,5,7,22,36,54,55,57,58,59,60,62,63,64,65,66,67,68,69,37,39,40,41,42,44,45,46,47,48,49,50,52,8,10,11,12,13,15,16,17,18,19,20,23,25,26,27,28,30,31,32,33,34],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\base.py":[1,2,3,4,7,8,13,16,31,9,10,11,17,18,19,20,21,22,23,25,26,27,28,29,14],"C:\\Users\\MARCUS\\Desktop\\CHALLENGE 3\\IREPORTER-CHALLENGE-3\\restapi\\tests\\test_user_endpoints.py":[1,2,4,6,14,32,52,72,7,8,9,11,12,54,55,56,57,58,59,60,62,63,65,66,68,69,70,17,18,19,20,21,22,23,25,26,27,28,29,30,74,75,76,77,78,79,80,82,83,85,86,87,88,89,90,91,93,94,95,96,97,98,34,35,36,37,38,39,40,42,43,45,46,48,49,50]}}
