+++ .gitignore
@@ -1,3 +1,4 @@
 venv/
 .vscode
 __pycache__/
+/htmlcov
